django_requirements.txt
django-cors-headers == 3.7.0
django-filter == 2.4.0
django-rest-swagger == 2.2.0
djangorestframework == 3.12.4
django-crispy-forms == 1.11.2